44 commits. Every fix has its guard committed red in the commit before it, with
the failure message quoted; where a guard could not be committed red -- because
expressing it needed a constructor the change itself introduced -- teeth were
verified by mutation and the mutation is quoted instead.

web/storage and web/scheduling had no test source at all and ModPackControllerTest
mocks its service, so everything below ModPackService was executed by nothing.
That is the single fact the pass turns on: the defects were not subtle, they were
unobserved.

What was fixed, worst first: the generation queue died permanently on the first
exception a task threw; the Nekodetector scan never ran for a ZIP, i.e. for the
entire web service, while logging that it did; uploads were stored before being
validated, so a rejected one cost the same three writes as an accepted one and
its GridFS copy could never be reclaimed; server-icon detection worked only on
modpacks WITHOUT a manifest, so every CurseForge and MultiMC export silently lost
it; a complete Modrinth parser was reachable from nothing; and any pack under
1 MiB was undownloadable, because size was truncated mebibytes and both tables
gated their download button on size > 0.

Two findings changed on contact with evidence and are recorded as such: the
upload filename was NOT an exploitable arbitrary write (the "-orig-" prefix makes
the first path component a directory that does not exist, so the open fails with
ENOENT -- measured against a real Tomcat), and the GridFS _id query was NOT broken
(MongoConverter.convertId converts a valid 24-hex String). Both had real defects
beside them, which is why they still needed fixing.

The suite then got a real database. H2 cannot work -- the driver speaks the
MongoDB wire protocol -- but flapdoodle runs a real mongod in-process, no Docker,
which closed the end-to-end verification Docker's kernel incompatibility had
blocked. Bounding the server-selection timeout took the app suite from 147.6s to
28.2s, and WebPersistenceIT now proves against a real server what was previously
only argued: the index exists, an upload round-trips both storage tiers, delete
reclaims both copies, and a refused duplicate writes none.

Full build green: 1986 JVM tests across six modules, 0 failures, plus 35 frontend
tests. Deferred with reasons in BACKLOG as B41 and B42.

+++ serverpackcreator-app/src/main/kotlin/de/griefed/serverpackcreator/app/web/serverpack/ServerPack.kt
@@ -42,7 +42,7 @@ class ServerPack {
     /** Which [ModPack][de.griefed.serverpackcreator.app.web.modpack.ModPack] this was generated from. */
     var modpackId: String = ""
     /** Archive size in bytes. */
-    var size: Int = 0
+    var size: Long = 0
     /** How often this server pack has been downloaded. */
     var downloads: Int = 0
     /** How many users have voted that this pack actually runs — a counter, not a flag. */
@@ -53,7 +53,13 @@ class ServerPack {
     var fileID: String? = null
     /** The name the archive is served under, which need not match [fileID]. */
     var fileName: String? = null
-    /** SHA256 of the archive. Indexed, because it is what the de-duplication looks up. */
+    /**
+     * SHA256 of the archive.
+     *
+     * **Not** indexed, and nothing looks a server pack up by it — the de-duplication this doc used to
+     * claim exists only for modpacks, on `ModPack.sha256`, which does carry `@Indexed`. Kept because it
+     * is served to clients that want to verify a download.
+     */
     var sha256: String? = null
 
     /** The configuration this pack was generated with — what makes two packs from one modpack different. */
@@ -61,7 +67,7 @@ class ServerPack {
     var runConfiguration: RunConfiguration? = null
 
     constructor(
-        size: Int,
+        size: Long,
         runConfiguration: RunConfiguration?,
         fileID: String?,
         fileName: String?,
@@ -79,7 +85,7 @@ class ServerPack {
     @PersistenceCreator
     private constructor(
         id: String,
-        size: Int,
+        size: Long,
         downloads: Int,
         confirmedWorking: Int,
         dateCreated: Date,
